refactor(shell): extract handleInputEvent into AgentModeSession (stage C)
Move section 4 of the agent loop — the per-event input dispatch (the four
QuestionComponent step()-based inline prompts, the streaming and non-streaming
AgentInputComponent switches, the submit/plan-mode/model-switch handling) —
verbatim into AgentModeSession::handleInputEvent, returning LoopControl so the
agent-mode exit flows out cleanly: the Abort case's co_return becomes
LoopControl::Exit and the loop's continue statements become LoopControl::Continue.
runAgentModeFlow now does `if (session.handleInputEvent(...) == Exit) co_return;`.

To make this possible the session now also borrows the stable per-run
collaborators (worker, provider [by ref so model switches stay visible],
sessionManager, toolRegistry, mcpServerManager, history/file-path providers,
contextFuture, cwd) via its constructor. The duplicated system-prompt handshake
(initial-message path and submit path) is unified into
AgentModeSession::ensureSystemPromptReady(). AgentContextResult moves to a new
shell/AgentContext.hpp so the future type is nameable from the session.

Pure structural move, no behavior change. Verified: builds clean under
clang-tidy/ASAN, test-endo-shell/test-tui pass, and a tmux smoke test of agent
mode (enter via '#', type/clear input, Esc-confirm abort back to the shell
prompt) works without crashes. runAgentModeFlow statement proxy 790 -> 503.
