+++ src/shell/AgentModeSession.cpp
@@ -8,10 +8,19 @@
 #include <tui/Theme.hpp>
 
 #include <agent/AgentConfig.hpp>
+#include <agent/commands/AgentHistoryProvider.hpp>
+#include <agent/commands/FilePathCompleter.hpp>
+#include <agent/commands/SlashCommand.hpp>
+#include <agent/commands/SlashCommandRegistry.hpp>
+#include <agent/context/FileReferenceExpander.hpp>
+#include <agent/conversation/SessionManager.hpp>
 #include <agent/providers/ProviderFactory.hpp>
 #include <agent/providers/ProviderModels.hpp>
 #include <agent/session/AgentSession.hpp>
+#include <agent/session/AgentWorker.hpp>
 #include <agent/tools/DiffRenderer.hpp>
+#include <agent/tools/ExploreTool.hpp>
+#include <agent/tools/ToolRegistry.hpp>
 #include <agent/tracing/TraceTerminalRenderer.hpp>
 
 namespace endo
@@ -443,5 +452,623 @@ void AgentModeSession::drainAgentMessages(std::vector<agent::FromAgentMessage>& 
     // Suppress while ask-user is active — only one inline prompt at a time.
     if (streaming && !streamingPromptVisible && !anyPromptActive())
         renderStreamingPrompt();
+}
+
+void AgentModeSession::ensureSystemPromptReady()
+{
+    if (systemPromptReady)
+        return;
+    auto result = _contextFuture.get();
+    _shell._agentSession->setSystemPrompt(std::move(result.systemPrompt));
+    if (auto* explore = dynamic_cast<agent::ExploreTool*>(_toolRegistry.findTool("explore")))
+        explore->setSystemPrompt(std::move(result.exploreSystemPrompt));
+    if (!result.gitBranch.empty())
+        _inputComponent.setGitBranch(std::move(result.gitBranch));
+    if (!result.projectPath.empty())
+        _inputComponent.setProjectPath(std::move(result.projectPath));
+    _filePathProviderPtr->setFilePaths(result.projectContext.filePaths);
+    _shell._cachedProjectContext = std::move(result.projectContext);
+    _shell._cachedProjectContextCwd = _cwd;
+    systemPromptReady = true;
+}
+
+LoopControl AgentModeSession::handleInputEvent(
+    tui::InputEvent const& event,
+    bool& needsRedraw,
+    agent::SlashCommandRegistry const& slashRegistry,
+    std::function<bool(std::string_view, std::string_view)> const& switchToModel,
+    std::function<void()> const& saveHistory)
+{
+    // Aliases map the moved dispatch body's bare collaborator names onto the
+    // session's borrowed references, so the body below is otherwise verbatim.
+    auto& out = _out;
+    auto& terminal = _terminal;
+    auto& screen = _screen;
+    auto& inputComponent = _inputComponent;
+    auto& worker = _worker;
+    auto*& provider = _provider;
+    auto const& sessionManager = _sessionManager;
+    auto& mcpServerManager = _mcpServerManager;
+    auto* const historyProviderPtr = _historyProviderPtr;
+    auto const& theme = tui::currentTheme();
+
+    // During ask-user, route input to the question component.
+    if (askUserPrompt.active && askUserPrompt.component)
+    {
+        // Drive the question through its modal step(); a returned result
+        // means confirmed/cancelled, std::nullopt means continue (redraw
+        // only when the step changed visible state).
+        auto const stepResult = askUserPrompt.component->step(event);
+        if (stepResult && stepResult->confirmed)
+        {
+            {
+                auto const& answerText = stepResult->answer;
+                auto const qConfig = askUserPrompt.component->config();
+                auto const selectedIdx = stepResult->selectedIndex;
+                auto const& checkedIdx = stepResult->checkedIndices;
+                auto const otherActive = stepResult->otherActive;
+                worker.inbound().push(
+                    agent::UserAnswerMessage { .requestId = askUserPrompt.requestId,
+                                               .answer = agent::UserAnswer { .answer = answerText } });
+                askUserPrompt.clear(out);
+                // Echo question + options with selection to scrollback
+                {
+                    auto const barStyle = tui::Style { .fg = theme.agentColors.leftBar };
+                    auto const questionStyle = tui::Style { .fg = theme.colors.text };
+                    auto const normalStyle = tui::Style { .fg = theme.agentColors.statusText, .dim = true };
+                    auto const selectedStyle = tui::Style { .fg = theme.agentColors.leftBar, .bold = true };
+                    // Question text
+                    out.writeText("\u2502 ", barStyle);
+                    out.writeText(qConfig.questionText, questionStyle);
+                    out.linefeed();
+                    // Options
+                    if (qConfig.multiSelect)
+                    {
+                        auto const checkedSet = std::set<std::size_t>(checkedIdx.begin(), checkedIdx.end());
+                        for (auto i = std::size_t { 0 }; i < qConfig.options.size(); ++i)
+                        {
+                            auto const checked = checkedSet.contains(i);
+                            out.writeText("\u2502 ", barStyle);
+                            if (checked)
+                            {
+                                out.writeText(" \xe2\x96\xb6 " + qConfig.options[i], selectedStyle);
+                            }
+                            else
+                            {
+                                out.writeText("   " + qConfig.options[i], normalStyle);
+                            }
+                            out.linefeed();
+                        }
+                    }
+                    else
+                    {
+                        for (auto i = std::size_t { 0 }; i < qConfig.options.size(); ++i)
+                        {
+                            out.writeText("\u2502 ", barStyle);
+                            if (!otherActive && i == selectedIdx)
+                            {
+                                out.writeText(" \xe2\x96\xb6 " + qConfig.options[i], selectedStyle);
+                            }
+                            else
+                            {
+                                out.writeText("   " + qConfig.options[i], normalStyle);
+                            }
+                            out.linefeed();
+                        }
+                    }
+                    // Custom "Other..." text
+                    if (otherActive)
+                    {
+                        out.writeText("\u2502 ", barStyle);
+                        out.writeText(" \xe2\x96\xb6 " + answerText, selectedStyle);
+                        out.linefeed();
+                    }
+                    out.writeText("\u2502", barStyle);
+                    out.linefeed();
+                    out.flush();
+                }
+                askUserPrompt.reset();
+            }
+        }
+        else if (stepResult)
+        {
+            {
+                auto const qConfig = askUserPrompt.component->config();
+                worker.inbound().push(
+                    agent::UserAnswerMessage { .requestId = askUserPrompt.requestId,
+                                               .answer = agent::UserAnswer { .cancelled = true } });
+                askUserPrompt.clear(out);
+                // Echo question + options with cancellation to scrollback
+                {
+                    auto const barStyle = tui::Style { .fg = theme.agentColors.leftBar };
+                    auto const questionStyle = tui::Style { .fg = theme.colors.text };
+                    auto const normalStyle = tui::Style { .fg = theme.agentColors.statusText, .dim = true };
+                    auto const dimStyle = tui::Style { .fg = theme.agentColors.statusText, .dim = true };
+                    // Question text
+                    out.writeText("\u2502 ", barStyle);
+                    out.writeText(qConfig.questionText, questionStyle);
+                    out.linefeed();
+                    // Options (all unselected)
+                    for (auto const& opt: qConfig.options)
+                    {
+                        out.writeText("\u2502 ", barStyle);
+                        out.writeText("   " + opt, normalStyle);
+                        out.linefeed();
+                    }
+                    // Cancellation notice
+                    out.writeText("\u2502 ", barStyle);
+                    out.writeText(" (cancelled)", dimStyle);
+                    out.linefeed();
+                    out.writeText("\u2502", barStyle);
+                    out.linefeed();
+                    out.flush();
+                }
+                askUserPrompt.reset();
+            }
+        }
+        else if (askUserPrompt.component->stepChangedState())
+        {
+            auto guard = out.syncGuard();
+            askUserPrompt.clear(out);
+            askUserPrompt.render(out, terminal);
+        }
+        return LoopControl::Continue;
+    }
+
+    // During permission prompt, route input to the permission component.
+    if (permissionPrompt.active && permissionPrompt.component)
+    {
+        auto const stepResult = permissionPrompt.component->step(event);
+        if (stepResult && stepResult->confirmed)
+        {
+            {
+                auto const selectedIdx = stepResult->selectedIndex;
+                permissionPrompt.clear(out);
+
+                auto decision = agent::PermissionDecision::Denied;
+                if (selectedIdx == 0) // "Yes"
+                    decision = agent::PermissionDecision::Approved;
+                else if (selectedIdx == 1) // "Yes, always for this tool"
+                    decision = agent::PermissionDecision::Approved;
+                else // "No"
+                    decision = agent::PermissionDecision::Denied;
+
+                worker.inbound().push(agent::PermissionResponseMessage {
+                    .requestId = permissionPrompt.requestId,
+                    .decision = decision,
+                });
+
+                // Echo the permission decision to scrollback.
+                {
+                    auto const barStyle = tui::Style { .fg = theme.agentColors.leftBar };
+                    auto const dimStyle = tui::Style { .fg = theme.agentColors.statusText };
+                    out.writeText("\u2502 ", barStyle);
+                    if (decision == agent::PermissionDecision::Approved)
+                        out.writeText("Approved", dimStyle);
+                    else
+                        out.writeText("Denied", dimStyle);
+                    out.linefeed();
+                    out.flush();
+                }
+
+                permissionPrompt.reset();
+            }
+        }
+        else if (stepResult)
+        {
+            permissionPrompt.clear(out);
+            worker.inbound().push(agent::PermissionResponseMessage {
+                .requestId = permissionPrompt.requestId,
+                .decision = agent::PermissionDecision::Cancelled,
+            });
+
+            auto const barStyle = tui::Style { .fg = theme.agentColors.leftBar };
+            auto const dimStyle = tui::Style { .fg = theme.agentColors.statusText };
+            out.writeText("\u2502 ", barStyle);
+            out.writeText("(cancelled)", dimStyle);
+            out.linefeed();
+            out.flush();
+
+            permissionPrompt.reset();
+        }
+        else if (permissionPrompt.component->stepChangedState())
+        {
+            auto guard = out.syncGuard();
+            permissionPrompt.clear(out);
+            permissionPrompt.render(out, terminal);
+        }
+        return LoopControl::Continue;
+    }
+
+    // During session picker, route input to the session picker component.
+    if (sessionPickerPrompt.active && sessionPickerPrompt.component)
+    {
+        auto const stepResult = sessionPickerPrompt.component->step(event);
+        if (stepResult && stepResult->confirmed)
+        {
+            {
+                auto const selectedIdx = stepResult->selectedIndex;
+                sessionPickerPrompt.clear(out);
+                if (selectedIdx < sessionPickerNames.size())
+                {
+                    auto const& name = sessionPickerNames[selectedIdx];
+                    auto loaded = sessionManager.loadSession(name);
+                    if (loaded.has_value())
+                    {
+                        auto& [meta, messages] = *loaded;
+                        (*_shell._agentSession).reset();
+                        historyProviderPtr->setEntries({});
+                        for (auto const& msg: messages)
+                        {
+                            if (msg.role == agent::Role::User)
+                            {
+                                auto const text =
+                                    agent::FileReferenceExpander::stripExpansions(msg.textContent());
+                                if (!text.empty())
+                                    historyProviderPtr->addEntry(text);
+                            }
+                        }
+                        _shell._agentSession->loadPersistedMessages(std::move(messages));
+                        _shell._activeSessionName = name;
+                        _shell._sessionCreatedAt = meta.createdAt;
+                        sessionManager.setLastActiveSession(name);
+                        out.writeText("Session '" + name + "' loaded.\n");
+                    }
+                    else
+                    {
+                        auto const errorStyle = tui::Style { .fg = theme.agentColors.errorText };
+                        out.writeText("Failed to load session: " + loaded.error().message + "\n", errorStyle);
+                    }
+                }
+                sessionPickerPrompt.reset();
+                sessionPickerNames.clear();
+                out.flush();
+            }
+        }
+        else if (stepResult)
+        {
+            sessionPickerPrompt.clear(out);
+            sessionPickerPrompt.reset();
+            sessionPickerNames.clear();
+        }
+        else if (sessionPickerPrompt.component->stepChangedState())
+        {
+            auto guard = out.syncGuard();
+            sessionPickerPrompt.clear(out);
+            sessionPickerPrompt.render(out, terminal);
+        }
+        return LoopControl::Continue;
+    }
+
+    // During plan approval, route input to the approval component.
+    if (planApprovalPrompt.isActive())
+    {
+        auto const stepResult = planApprovalPrompt.component->step(event);
+        if (stepResult && stepResult->confirmed)
+        {
+            {
+                auto const selectedIdx = stepResult->selectedIndex;
+                planApprovalPrompt.clear(out);
+                planApprovalPrompt.reset();
+                if (selectedIdx == 0) // "Yes, execute"
+                {
+                    worker.inbound().push(agent::PlanApproveMessage { .plan = std::move(*pendingPlan) });
+                    pendingPlan.reset();
+                    streaming = true;
+                    inputComponent.setThinkingActive(true);
+                    inputComponent.setActivityLabel("Executing plan...");
+                }
+                else if (selectedIdx == 1) // "Yes, compact context first"
+                {
+                    worker.inbound().push(
+                        agent::PlanApproveMessage { .plan = std::move(*pendingPlan), .compactFirst = true });
+                    pendingPlan.reset();
+                    streaming = true;
+                    inputComponent.setThinkingActive(true);
+                    inputComponent.setActivityLabel("Compacting context...");
+                }
+                else if (selectedIdx == 2) // "No, discard"
+                {
+                    pendingPlan.reset();
+                    auto const dimStyle = tui::Style { .fg = theme.agentColors.statusText };
+                    out.writeText("Plan discarded.\n", dimStyle);
+                    out.flush();
+                    screen.releaseCursor();
+                    auto const newPrefSize = inputComponent.preferredSize();
+                    inputComponent.setArea(tui::Rect {
+                        .x = 0, .y = 0, .width = terminal.columns(), .height = newPrefSize.height });
+                    screen.draw();
+                }
+                else // "Revise"
+                {
+                    pendingPlan.reset();
+                    // Stay in plan mode for revision.
+                    screen.releaseCursor();
+                    auto const newPrefSize = inputComponent.preferredSize();
+                    inputComponent.setArea(tui::Rect {
+                        .x = 0, .y = 0, .width = terminal.columns(), .height = newPrefSize.height });
+                    screen.draw();
+                }
+            }
+        }
+        else if (stepResult)
+        {
+            planApprovalPrompt.clear(out);
+            planApprovalPrompt.reset();
+            pendingPlan.reset();
+            {
+                screen.releaseCursor();
+                auto const newPrefSize = inputComponent.preferredSize();
+                inputComponent.setArea(
+                    tui::Rect { .x = 0, .y = 0, .width = terminal.columns(), .height = newPrefSize.height });
+                screen.draw();
+            }
+        }
+        else if (planApprovalPrompt.component->stepChangedState())
+        {
+            auto guard = out.syncGuard();
+            planApprovalPrompt.clear(out);
+            planApprovalPrompt.render(out, terminal);
+        }
+        return LoopControl::Continue;
+    }
+
+    // During streaming, only handle Escape (cancel) and Ctrl+L (clear).
+    if (streaming)
+    {
+        auto const action = inputComponent.processInput(event);
+        switch (action)
+        {
+            case agent::AgentInputComponent::Action::Abort:
+                streamCancelled = true;
+                worker.inbound().push(agent::CancelMessage {});
+                break;
+            case agent::AgentInputComponent::Action::ClearScreen:
+                out.clearScreen();
+                out.flush();
+                streamingPromptVisible = false;
+                renderStreamingPrompt();
+                break;
+            case agent::AgentInputComponent::Action::Changed: {
+                auto guard = out.syncGuard();
+                clearStreamingPrompt();
+                renderStreamingPrompt();
+                break;
+            }
+            default: break;
+        }
+        return LoopControl::Continue;
+    }
+
+    // Not streaming — handle full input.
+    auto const action = inputComponent.processInput(event);
+    switch (action)
+    {
+        case agent::AgentInputComponent::Action::Submit: {
+            // Poll MCP servers for tool list changes before each LLM turn.
+            mcpServerManager.processNotifications();
+
+            auto sentToWorker = false;
+
+            // Ensure system prompt is ready.
+            ensureSystemPromptReady();
+
+            auto const query = std::string(inputComponent.text());
+
+            // Extract attached images before clearing.
+            auto attachedImages = std::vector<agent::ImageBlock>(inputComponent.attachedImages().begin(),
+                                                                 inputComponent.attachedImages().end());
+
+            // Expand @-file references for agent context injection.
+            auto const expandFileRefs = [&](std::string_view text) {
+                return agent::FileReferenceExpander::expand(text, std::filesystem::current_path())
+                    .expandedMessage;
+            };
+
+            if (!query.starts_with("/"))
+            {
+                inputComponent.inputField().addHistory(query);
+                historyProviderPtr->addEntry(query);
+            }
+
+            // Move cursor past the input component, preserving image previews in scrollback.
+            auto const totalLines = inputComponent.inputField().lineCount();
+            auto const cursorLine = inputComponent.inputField().cursorLine();
+            auto const previewLines = inputComponent.imagePreviewHeight();
+            inputComponent.clear();
+            auto const linesToMoveDown = totalLines - cursorLine + previewLines;
+            if (linesToMoveDown > 0)
+                out.moveDown(linesToMoveDown);
+            out.carriageReturn();
+            out.clearLine(); // Clear info line (shortcut hints / spinner)
+            out.linefeed();
+            out.clearLine(); // Clear bottom padding (NBSP marker)
+            out.flush();
+
+            screen.releaseCursor();
+
+            // Dispatch slash commands.
+            if (query.starts_with("/"))
+            {
+                auto const spacePos = query.find(' ');
+                auto const cmdName =
+                    query.substr(1, spacePos == std::string::npos ? std::string::npos : spacePos - 1);
+                auto const args = spacePos != std::string::npos ? query.substr(spacePos + 1) : "";
+
+                if (auto const* cmd = slashRegistry.findCommand(cmdName))
+                {
+                    auto commandResult = cmd->execute(args);
+                    if (auto const* d = std::get_if<agent::DirectOutput>(&commandResult))
+                    {
+                        out.writeText(d->text);
+                        out.flush();
+                    }
+                    else if (auto const* m = std::get_if<agent::MarkdownOutput>(&commandResult))
+                    {
+                        auto mdRenderer = tui::MarkdownRenderer(out);
+                        mdRenderer.setMaxWidth(terminal.columns());
+                        mdRenderer.render(m->markdown);
+                        out.flush();
+                    }
+                    else if (auto const* p = std::get_if<agent::PlanModeRequest>(&commandResult))
+                    {
+                        if (!_shell.agentConfig.planMode.enabled)
+                        {
+                            auto const errorStyle = tui::Style { .fg = theme.agentColors.errorText };
+                            out.writeText("Plan mode is disabled in configuration.\n", errorStyle);
+                            out.flush();
+                        }
+                        else if (p->query.empty())
+                        {
+                            if (!planModeActive)
+                            {
+                                planModeActive = true;
+                                inputComponent.setPlanMode(true);
+                            }
+                            auto const infoStyle = tui::Style { .fg = theme.agentColors.statusText };
+                            out.writeText("Plan mode active. Type your task to generate a plan.\n",
+                                          infoStyle);
+                            out.flush();
+                        }
+                        else
+                        {
+                            // Send plan query to worker.
+                            worker.inbound().push(agent::UserPromptMessage { .text = expandFileRefs(p->query),
+                                                                             .planMode = true });
+                            sentToWorker = true;
+                        }
+                    }
+                    else if (auto const* r = std::get_if<agent::PromptRewrite>(&commandResult))
+                    {
+                        // Send rewritten prompt to worker.
+                        worker.inbound().push(agent::UserPromptMessage { .text = expandFileRefs(r->prompt) });
+                        sentToWorker = true;
+                    }
+                    else if (auto const* sp = std::get_if<agent::SessionPickerRequest>(&commandResult))
+                    {
+                        // Show interactive session picker using QuestionComponent.
+                        sessionPickerNames = sp->sessionNames;
+                        sessionPickerPrompt.component.emplace(tui::QuestionConfig {
+                            .questionText = sp->questionText,
+                            .options = sp->options,
+                            .multiSelect = false,
+                            .allowOther = false,
+                        });
+                        sessionPickerPrompt.active = true;
+                        sessionPickerPrompt.render(out, terminal);
+                    }
+                }
+                else
+                {
+                    auto const errorStyle = tui::Style { .fg = theme.agentColors.errorText };
+                    out.writeText("Unknown command: /" + cmdName + "\n", errorStyle);
+                    out.flush();
+                }
+            }
+            else if (planModeActive && _shell.agentConfig.planMode.enabled)
+            {
+                // Plan mode: send to worker with planMode flag.
+                worker.inbound().push(agent::UserPromptMessage {
+                    .text = expandFileRefs(query),
+                    .planMode = true,
+                    .images = std::move(attachedImages),
+                });
+                sentToWorker = true;
+                saveHistory();
+            }
+            else
+            {
+                // Normal message: send to worker.
+                worker.inbound().push(agent::UserPromptMessage {
+                    .text = expandFileRefs(query),
+                    .images = std::move(attachedImages),
+                });
+                sentToWorker = true;
+            }
+
+            // Re-render input component only for non-streaming commands.
+            // Streaming responses re-render via CompletionMessage handler.
+            if (!sentToWorker)
+            {
+                auto const newPrefSize = inputComponent.preferredSize();
+                inputComponent.setArea(
+                    tui::Rect { .x = 0, .y = 0, .width = terminal.columns(), .height = newPrefSize.height });
+                screen.draw();
+            }
+            break;
+        }
+        case agent::AgentInputComponent::Action::Abort: {
+            // Stop worker before exiting agent mode.
+            worker.stop();
+            _shell._agentSession->setPermissionManager(nullptr);
+            _shell._agentSession->setToolRegistry(nullptr);
+            _shell._agentSession->setToolStatusCallback(nullptr);
+            _shell._agentSession->setTracer(nullptr);
+            terminal.input().setWakeup(nullptr);
+            screen.clearAndRelease();
+            return LoopControl::Exit;
+        }
+        case agent::AgentInputComponent::Action::CycleMode: {
+            planModeActive = !planModeActive;
+            inputComponent.setPlanMode(planModeActive);
+            needsRedraw = true;
+            break;
+        }
+        case agent::AgentInputComponent::Action::CycleThinkingMode: {
+            // Cycle thinking mode for the active provider.
+            auto const& pName = _shell._agentProviderFactory->activeProviderName();
+            auto* thinkingModePtr = static_cast<agent::ThinkingMode*>(nullptr);
+            if (pName == "claude")
+                thinkingModePtr = &_shell.agentConfig.claude.thinkingMode;
+            else if (pName == "openai")
+                thinkingModePtr = &_shell.agentConfig.openai.thinkingMode;
+            else if (pName == "openai_compat")
+                thinkingModePtr = &_shell.agentConfig.openaiCompat.thinkingMode;
+            else if (pName == "gemini")
+                thinkingModePtr = &_shell.agentConfig.gemini.thinkingMode;
+
+            if (thinkingModePtr)
+            {
+                *thinkingModePtr = agent::nextThinkingMode(*thinkingModePtr);
+                inputComponent.setThinkingMode(*thinkingModePtr);
+                auto const currentModel = provider->modelInfo().modelName;
+                switchToModel(pName, currentModel);
+            }
+            needsRedraw = true;
+            break;
+        }
+        case agent::AgentInputComponent::Action::CycleModel: {
+            // Cycle through hardcoded model list for the active provider.
+            auto const& pName = _shell._agentProviderFactory->activeProviderName();
+            auto const models = agent::modelsForProvider(pName);
+            if (!models.empty())
+            {
+                auto const currentModel = provider->modelInfo().modelName;
+                auto const nextModelName = agent::nextModel(models, currentModel);
+                switchToModel(pName, nextModelName);
+            }
+            needsRedraw = true;
+            break;
+        }
+        case agent::AgentInputComponent::Action::ClearScreen: {
+            out.clearScreen();
+            out.flush();
+            screen.releaseCursor();
+            needsRedraw = true;
+            break;
+        }
+        case agent::AgentInputComponent::Action::CommandPalette:
+            // Palette is shown internally by AgentInputComponent; just redraw.
+            needsRedraw = true;
+            break;
+        case agent::AgentInputComponent::Action::NewPrompt:
+            inputComponent.clear();
+            needsRedraw = true;
+            break;
+        case agent::AgentInputComponent::Action::Changed: needsRedraw = true; break;
+        case agent::AgentInputComponent::Action::None: break;
+    }
+
+    return LoopControl::Continue;
 }
 } // namespace endo
